+++ src/django_spicy_id/tests/models.py
@@ -1,6 +1,18 @@
+from django.core.exceptions import ValidationError
 from django.db import models
 
-from django_spicy_id import SpicyAutoField, SpicyBigAutoField, SpicyUUIDField
+from django_spicy_id import (
+    SpicyAutoField,
+    SpicyBigAutoField,
+    SpicyUUIDField,
+    TypeIDField,
+)
+
+
+def reject_thirteen(value):
+    """Test validator: rejects the id "ex_D" (13 in base62)."""
+    if value == "ex_D":
+        raise ValidationError("thirteen is unlucky", code="unlucky")
 
 
 class Model_WithDefaults(models.Model):
@@ -47,3 +59,15 @@ class UUIDModel_Hex(models.Model):
 
 class UUIDModel_Base58(models.Model):
     id = SpicyUUIDField("uu", primary_key=True, encoding="b58")
+
+
+class Model_WithCustomValidator(models.Model):
+    id = SpicyBigAutoField("ex", primary_key=True, validators=[reject_thirteen])
+
+
+class TypeIDModel(models.Model):
+    id = TypeIDField("prefix", primary_key=True)
+
+
+class TypeIDModel_NoPrefix(models.Model):
+    id = TypeIDField(primary_key=True)